+++ tlRender/lib/tlTimelineUI/TimelineItem.cpp
@@ -7,6 +7,7 @@
 #include <tlTimelineUI/AudioClipItem.h>
 #include <tlTimelineUI/GapItem.h>
 #include <tlTimelineUI/VideoClipItem.h>
+#include <tlTimelineUI/EffectItem.h>
 
 #include <tlUI/DrawUtil.h>
 #include <tlUI/ScrollArea.h>
@@ -89,6 +90,7 @@ namespace tl
                     track.durationLabel->setMarginRole(
                         ui::SizeRole::MarginInside);
 
+                    otio::Item* item = nullptr;
                     for (const auto& child : otioTrack->children())
                     {
                         if (auto clip =
@@ -112,6 +114,24 @@ namespace tl
                                 break;
                             }
                             track.otioIndexes.push_back(otioIndex);
+
+                            auto effects = clip->effects();
+                            std::vector<std::shared_ptr<EffectItem> > effectItems;
+
+                            otio::Item* item = clip;
+                            
+                            effectItems.reserve(effects.size());
+                            for (const auto effect : effects)
+                            {
+                                effectItems.push_back(EffectItem::create(
+                                                          effect, item, scale, options, displayOptions,
+                                                          itemData, context, shared_from_this()));
+                            }
+
+                            if (!effectItems.empty())
+                            {
+                                track.effects[otioIndex] = effectItems;
+                            }
                         }
                         else if (
                             auto gap =
@@ -301,6 +321,31 @@ namespace tl
                                 _scale,
                         y + std::max(labelSizeHint.h, durationSizeHint.h),
                         sizeHint.w, track.clipHeight));
+
+                    for (auto index : track.otioIndexes)
+                    {
+                        auto i = track.effects.find(index);
+                        if (i == track.effects.end())
+                            continue;
+
+                        auto effects = i->second;
+                        int effectsY = y;
+                        for (const auto& effect : effects)
+                        {
+                            const math::Size2i& sizeHint = effect->getSizeHint();
+                        
+                            const otime::TimeRange& timeRange =
+                                effect->getTimeRange();
+                            effect->setGeometry(math::Box2i(
+                                                    _geometry.min.x + timeRange.start_time()
+                                                    .rescaled_to(1.0)
+                                                    .value() *
+                                                    _scale,
+                                                    effectsY, sizeHint.w, sizeHint.h));
+
+                            effectsY += sizeHint.h;
+                        }
+                    }
                 }
 
                 if (visible)