+++ tlRender/lib/tlTimelineUI/TimelineItemRoll.cpp
@@ -6,6 +6,8 @@
 #include <tlTimelineUI/TimelineItem.h>
 #include <tlTimelineUI/TimelineItemPrivate.h>
 
+#include <opentimelineio/gap.h>
+
 namespace tl
 {
     namespace timelineui
@@ -22,7 +24,87 @@ namespace tl
                 break;
             case Private::MouseMode::Transition:
             {
-                _mouseMoveEventTrim(event);                
+                _mouse.pos.y = _mouse.pressPos.y;
+                const int offset = _mouse.pos.x - _mouse.pressPos.x;
+                math::Box2i move;
+                    
+                for (const auto& item : p.mouse.items)
+                {
+                    const math::Box2i& g = item->geometry;
+                    auto transitionItem = static_cast<TransitionItem*>(item->p.get());
+                    otime::TimeRange timeRange = item->p->getTimeRange();
+
+                    // Get transition items (ie. the clips associated to the
+                    // transition) time ranges.
+                    const int transitionTrack = item->track;
+                    std::vector<otime::TimeRange> itemRanges;
+                    _getTransitionTimeRanges(itemRanges, transitionTrack,
+                                             timeRange);
+                        
+                    if (p.mouse.side == Private::MouseClick::Left)
+                    {
+                        math::Size2i size = g.getSize();
+                        size.w -= offset;
+                        
+                        move = math::Box2i(
+                            g.min + _mouse.pos - _mouse.pressPos,
+                            size );
+                    }
+                    else
+                    {
+                        math::Size2i size = g.getSize();
+                        size.w += offset;
+                        move = math::Box2i(g.min, size);
+                    }
+
+                    const otime::RationalTime& startTime = posToTime(move.x());
+                    const otime::RationalTime& duration  = posToTime(move.x() + move.w()) - startTime;
+                        
+                    timeRange = otime::TimeRange(startTime - _timeRange.start_time(), duration);
+                    const otime::RationalTime in_offset = itemRanges[1].start_time() -
+                                                          timeRange.start_time();
+                    const otime::RationalTime out_offset = timeRange.end_time_exclusive() -
+                                                           itemRanges[1].start_time();
+
+                    if (in_offset.value() <= 1.F ||
+                        out_offset.value() <= 1.F)
+                        continue;
+
+                    // Clamp on clips.
+                    if (duration.value() <= 2.F)
+                    {
+                        continue;
+                    }
+
+                    if (timeRange.start_time() <= itemRanges[0].start_time())
+                    {
+                        continue;
+                    }
+                    
+                    if (timeRange.start_time() >= itemRanges[0].end_time_exclusive())
+                    {
+                        continue;
+                    }
+                    
+                    if (timeRange.end_time_exclusive() >= itemRanges[1].end_time_inclusive())
+                    {
+                        continue;
+                    }
+                    
+                    if (timeRange.end_time_exclusive() <= itemRanges[0].end_time_inclusive())
+                    {
+                        continue;
+                    }
+                    
+                    // Clamp on other transitions.
+                    if (_transitionIntersects(item->p, transitionTrack, timeRange))
+                    {
+                        continue;
+                    }
+                    
+                    transitionItem->setDurationLabel(std::to_string(int(duration.value())));
+                    item->p->setGeometry(move);
+                }
                 break;
             }
             case Private::MouseMode::Item:
@@ -80,6 +162,8 @@ namespace tl
                 return;
             switch (p.mouse.mode)
             {
+            case Private::MouseMode::kNone:
+                break;
             case Private::MouseMode::CurrentTime:
             {
                 break;
@@ -89,32 +173,35 @@ namespace tl
                 _mouse.pos.y = _mouse.pressPos.y;
                 const int offset = _mouse.pos.x - _mouse.pressPos.x;
                 math::Box2i move;
-                
                 std::vector<timeline::MoveData> moveData;
+                
+                const auto otioTimeline = p.player->getTimeline()->getTimeline();
                 moveData.push_back(
                     {
                         timeline::MoveType::UndoOnly
                     });
                 if (p.moveCallback)
                     p.moveCallback(moveData);
-                
-                const auto otioTimeline = p.player->getTimeline()->getTimeline();
+
                 for (const auto& item : p.mouse.items)
                 {
                     const math::Box2i& g = item->geometry;
                     otime::TimeRange timeRange = item->p->getTimeRange();
                     otime::TimeRange origRange = timeRange;
                                         
-                    math::Size2i size = g.getSize();
                     if (p.mouse.side == Private::MouseClick::Left)
                     {
+                        math::Size2i size = g.getSize();
                         size.w -= offset;
+                        
                         move = math::Box2i(
                             g.min + _mouse.pos - _mouse.pressPos,
-                            size);
+                            size );
+
                     }
                     else if (p.mouse.side == Private::MouseClick::Right)
                     {
+                        math::Size2i size = g.getSize();
                         size.w += offset;
                         move = math::Box2i(g.min, size);
                     }
@@ -125,9 +212,6 @@ namespace tl
 
                     startTime -= _timeRange.start_time();
                     timeRange = otime::TimeRange(startTime, duration);
-#if DEBUG_TIME_RANGES
-                    std::cerr << __LINE__ << " shifted time range=" << timeRange << std::endl;
-#endif
 
                     // Clamp on clips.
                     if (duration.value() < 1.F)
@@ -153,6 +237,27 @@ namespace tl
                             timeRange = otime::TimeRange(origRange.start_time() + startOffset,
                                                          origRange.duration() + durationOffset);
                             otioItem->set_source_range(timeRange);
+
+                            otime::TimeRange gapRange;
+                            int otioGapIndex = 0;
+                            if (p.mouse.side == Private::MouseClick::Right)
+                            {
+                                gapRange = otime::TimeRange(
+                                    otime::RationalTime(0,
+                                                        timeRange.duration().rate()),
+                                    -durationOffset);
+                                otioGapIndex = otioItemIndex + 1;
+                            }
+                            else
+                            {
+                                gapRange = otime::TimeRange(
+                                    otime::RationalTime(0,
+                                                        timeRange.duration().rate()),
+                                    -durationOffset);
+                                otioGapIndex = otioItemIndex;
+                            }
+                            otio::Gap* gap = new otio::Gap(gapRange);
+                            otioTrack->insert_child(otioGapIndex, gap);
                         }
                     }
                 }
@@ -161,7 +266,136 @@ namespace tl
             }
             case Private::MouseMode::Transition:
             {
-                _mouseReleaseEventTrim(event);
+                _mouse.pos.y = _mouse.pressPos.y;
+                const int offset = _mouse.pos.x - _mouse.pressPos.x;
+                math::Box2i move;
+                std::vector<timeline::MoveData> moveData;
+                for (const auto& item : p.mouse.items)
+                {
+                    const math::Box2i& g = item->geometry;
+                    otime::TimeRange timeRange = item->p->getTimeRange();
+                    
+                    // Get transition items (ie. the clips associated to the
+                    // transition) time ranges.
+                    const int transitionTrack = item->track;
+                    const int transitionIndex = item->index;
+                    std::vector<otime::TimeRange> itemRanges;
+                    _getTransitionTimeRanges(itemRanges, transitionTrack,
+                                             timeRange);
+                    
+                    const otime::RationalTime oneFrame =
+                        otime::RationalTime(1.F, timeRange.duration().rate());
+
+                    if (p.mouse.side == Private::MouseClick::Left)
+                    {
+                        math::Size2i size = g.getSize();
+                        size.w -= offset;
+                        
+                        move = math::Box2i(
+                            g.min + _mouse.pos - _mouse.pressPos,
+                            size );
+
+                    }
+                    else if (p.mouse.side == Private::MouseClick::Right)
+                    {
+                        math::Size2i size = g.getSize();
+                        size.w += offset;
+                        move = math::Box2i(g.min, size);
+                    }
+
+                    otime::RationalTime startTime = posToTime(move.x());
+                    otime::RationalTime duration  = posToTime(move.x() + move.w()) -
+                                                    startTime;
+
+                    startTime -= _timeRange.start_time();
+                    timeRange = otime::TimeRange(startTime, duration);
+
+                    // Clamp on clips.
+                    if (duration.value() <= 2.F)
+                    {
+                        duration  = otime::RationalTime(2.F, duration.rate());
+                        timeRange = otime::TimeRange(startTime, duration);
+                        
+                    }
+                    
+                    
+                    if (timeRange.start_time() <= itemRanges[0].start_time())
+                    {
+                        auto diff = itemRanges[0].start_time() - timeRange.start_time()
+                                    + oneFrame;
+                        timeRange = otime::TimeRange(itemRanges[0].start_time() + oneFrame,
+                                                     timeRange.duration() - diff);
+                    }
+                    
+                    if (timeRange.start_time() >= itemRanges[0].end_time_exclusive())
+                    {
+                        timeRange = otime::TimeRange::range_from_start_end_time_inclusive(
+                            itemRanges[1].start_time() - oneFrame,
+                            timeRange.end_time_inclusive());
+                    }
+                        
+                    if (timeRange.end_time_exclusive() >= itemRanges[1].end_time_inclusive())
+                    {
+                        timeRange = otime::TimeRange::range_from_start_end_time_inclusive(
+                            timeRange.start_time(),
+                            itemRanges[1].end_time_inclusive());
+                        timeRange = timeRange.duration_extended_by(-oneFrame);
+                    }
+                    
+                    if (timeRange.end_time_exclusive() <= itemRanges[0].end_time_inclusive())
+                    {
+                        timeRange = otime::TimeRange::range_from_start_end_time_inclusive(
+                            timeRange.start_time(),
+                            itemRanges[0].end_time_inclusive());
+                        timeRange = timeRange.duration_extended_by(oneFrame);
+                    }
+                    
+                    // Clamp on other transitions.
+                    if (_transitionIntersects(item->p, transitionTrack, timeRange))
+                    {
+                        for (const auto& transition : p.tracks[transitionTrack].transitions)
+                        {
+                            if (item->p == transition)
+                                continue;
+                            const otime::TimeRange transitionRange = transition->getTimeRange();
+                            if (timeRange.start_time() <= transitionRange.end_time_exclusive())
+                            {
+                                auto diff = transitionRange.end_time_inclusive() - timeRange.start_time();
+                                timeRange = otime::TimeRange(timeRange.start_time() + diff + oneFrame,
+                                                             timeRange.duration() - diff - oneFrame);
+
+                            }
+                            else if (timeRange.end_time_exclusive() >= transitionRange.start_time())
+                            {
+                                timeRange = otime::TimeRange(transitionRange.start_time() - oneFrame,
+                                                             timeRange.duration() + oneFrame);
+                            }
+                        }
+                    }
+                    
+                    const otime::RationalTime in_offset = itemRanges[1].start_time() -
+                                                          timeRange.start_time();
+                    const otime::RationalTime out_offset = timeRange.end_time_exclusive() -
+                                                           itemRanges[1].start_time();
+                    assert(timeRange.contains(itemRanges[1].start_time()));
+                    assert(in_offset.value() >= 1.F);
+                    assert(out_offset.value() >= 1.F);
+                    
+                    const int transitionOtioIndex =
+                        p.tracks[transitionTrack].otioTransitionIndexes[transitionIndex];
+                    moveData.push_back(
+                        {
+                            timeline::MoveType::Transition,
+                            transitionTrack, transitionIndex, transitionOtioIndex,
+                            transitionTrack, transitionIndex, transitionOtioIndex,
+                            in_offset, out_offset
+                        });
+                }
+                if (p.moveCallback)
+                    p.moveCallback(moveData);
+                auto otioTimeline = timeline::move(
+                    p.player->getTimeline()->getTimeline().value, moveData);
+                p.player->getTimeline()->setTimeline(otioTimeline);
             }
             }   
         }